# Maputil speedups