This is an attempt at something simpler than https://github.com/sunpy/sunpy/pull/4446 to speed up `contains_full_disk` (and other maputils functions). Times are for a 4k AIA image.

*`contains_full_disk`*
The bottleneck is running `~is_all_off_disk(smap)`, which is unnescessary. If the solar limb doesn't intersect the edges of the map (the first test), then the full disc is in the map iff (0, 0) is in the map. This is pretty easy to test for by just checking if both of the corner coordinates change sign, which is what this PR does.
11s > 0.3s

*`is_all_off_disk`*
Again, do not need to check if all the pixels are off disc, it's sufficient to check that the edges are off disk (ie. more than a solar radius away) and the solar center isn't in the image.
8s > 0.3s

*`is_all_on_disk`*
This is only true if all the edges are on the disc.
7s > 0.1s

*`contains_limb`*
Shortcut the case where the full disc is in the map. If not, it is sufficient to check if some edge pixels are both on and off the disc.
7s > 0.3s

A typical profile now looks like
```python
0.129 <module>  test.py:1
└─ 0.129 is_all_on_disk  sunpy/map/maputils.py:260
   ├─ 0.123 _edge_coordinates  sunpy/map/maputils.py:127
   │  ├─ 0.103 wrapper  astropy/units/decorators.py:165
   │  │  └─ 0.103 pixel_to_world  sunpy/map/mapbase.py:1099
   │  │     ├─ 0.049 coordinate_frame  sunpy/map/mapbase.py:482
   │  │     │  ├─ 0.042 wcs  sunpy/map/mapbase.py:421
   │  │     │  │  ├─ 0.015 __init__  astropy/wcs/wcs.py:359
   │  │     │  │  │     [31 frames hidden]  astropy
   │  │     │  │  ├─ 0.015 fits_header  sunpy/map/mapbase.py:983
   │  │     │  │  │  └─ 0.015 header_to_fits  sunpy/io/fits.py:194
   │  │     │  │  │     ├─ 0.007 __init__  astropy/io/fits/card.py:156
   │  │     │  │  │     │     [24 frames hidden]  astropy
   │  │     │  │  │     ├─ 0.004 append  astropy/io/fits/header.py:1157
   │  │     │  │  │     │     [5 frames hidden]  astropy
   │  │     │  │  │     └─ 0.003 [self]  
   │  │     │  │  └─ 0.010 observer_coordinate  sunpy/map/mapbase.py:830
   │  │     │  │     └─ 0.009 __getattr__  astropy/coordinates/sky_coordinate.py:577
   │  │     │  │           [32 frames hidden]  astropy
   │  │     │  │              0.008 __call__  astropy/coordinates/transformations.py:1208
   │  │     │  │              ├─ 0.004 wrapped_func  sunpy/coordinates/transformations.py:154
   │  │     │  │              │  └─ 0.004 hcrs_to_hgs  sunpy/coordinates/transformations.py:531
   │  │     │  └─ 0.006 wcs_to_celestial_frame  astropy/wcs/utils.py:182
   │  │     │        [3 frames hidden]  astropy
   │  │     │           0.005 solar_wcs_frame_mapping  sunpy/coordinates/wcs_utils.py:16
   │  │     │           └─ 0.005 __init__  sunpy/coordinates/frames.py:118
   │  │     │              └─ 0.005 __init__  astropy/coordinates/baseframe.py:415
   │  │     │                 └─ 0.005 __get__  sunpy/coordinates/frameattributes.py:129
   │  │     │                    └─ 0.004 __get__  astropy/coordinates/attributes.py:95
   │  │     │                       └─ 0.004 convert_input  sunpy/coordinates/frameattributes.py:100
   │  │     │                          └─ 0.004 convert_input  astropy/coordinates/attributes.py:415
   │  │     │                                [7 frames hidden]  astropy
   │  │     │                                   0.002 __call__  astropy/coordinates/transformations.py:913
   │  │     │                                   └─ 0.002 wrapped_func  sunpy/coordinates/transformations.py:154
   │  │     │                                      └─ 0.002 hgs_to_hgs  sunpy/coordinates/transformations.py:612
   │  │     ├─ 0.046 wcs  sunpy/map/mapbase.py:421
   │  │     │  ├─ 0.016 fits_header  sunpy/map/mapbase.py:983
   │  │     │  │  └─ 0.016 header_to_fits  sunpy/io/fits.py:194
   │  │     │  │     ├─ 0.008 __init__  astropy/io/fits/card.py:156
   │  │     │  │     │     [25 frames hidden]  astropy, <built-in>
   │  │     │  │     └─ 0.006 append  astropy/io/fits/header.py:1157
   │  │     │  │           [3 frames hidden]  astropy
   │  │     │  ├─ 0.014 __init__  astropy/wcs/wcs.py:359
   │  │     │  │     [37 frames hidden]  astropy, <built-in>
   │  │     │  ├─ 0.012 observer_coordinate  sunpy/map/mapbase.py:830
   │  │     │  │  └─ 0.010 __getattr__  astropy/coordinates/sky_coordinate.py:577
   │  │     │  │        [35 frames hidden]  astropy, <__array_function__ internal...
   │  │     │  │           0.009 __call__  astropy/coordinates/transformations.py:1208
   │  │     │  │           ├─ 0.004 wrapped_func  sunpy/coordinates/transformations.py:154
   │  │     │  │           │  └─ 0.004 hcrs_to_hgs  sunpy/coordinates/transformations.py:531
   │  │     │  │           │     └─ 0.002 _rotation_matrix_reprs_to_xz_about_z  sunpy/coordinates/transformations.py:491
   │  │     │  │           │        └─ 0.002 _rotation_matrix_reprs_to_reprs  sunpy/coordinates/transformations.py:470
   │  │     │  └─ 0.002 date  sunpy/map/mapbase.py:633
   │  │     │     └─ 0.002 parse_time  sunpy/time/time.py:253
   │  │     │        └─ 0.002 wrapper  functools.py:870
   │  │     │           └─ 0.002 convert_time_str  sunpy/time/time.py:198
   │  │     ├─ 0.004 __init__  astropy/coordinates/sky_coordinate.py:210
   │  │     │     [5 frames hidden]  astropy, <built-in>
   │  │     │        0.002 __init__  sunpy/coordinates/frames.py:118
   │  │     │        └─ 0.002 __init__  astropy/coordinates/baseframe.py:415
   │  │     │           └─ 0.002 __get__  sunpy/coordinates/frameattributes.py:129
   │  │     │              └─ 0.002 __get__  astropy/coordinates/attributes.py:95
   │  │     │                    [3 frames hidden]  astropy
   │  │     │        0.001 __get__  astropy/coordinates/attributes.py:95
   │  │     │        └─ 0.001 convert_input  sunpy/coordinates/frameattributes.py:32
   │  │     │           └─ 0.001 __eq__  astropy/time/core.py:2044
   │  │     │                 [7 frames hidden]  astropy
   │  │     └─ 0.004 wcs_pix2world  astropy/wcs/wcs.py:1386
   │  │           [6 frames hidden]  astropy, <built-in>
   │  └─ 0.020 map_edges  sunpy/map/maputils.py:57
   │     └─ 0.020 __rmul__  astropy/units/core.py:701
   │           [4 frames hidden]  astropy, <built-in>
   └─ 0.006 wrapper  astropy/units/decorators.py:165
      └─ 0.006 coordinate_is_on_solar_disk  sunpy/map/maputils.py:198
         ├─ 0.003 __getattr__  astropy/coordinates/sky_coordinate.py:577
         │     [2 frames hidden]  astropy
         │        0.003 __getattr__  astropy/coordinates/baseframe.py:1527
         │        └─ 0.003 represent_as  sunpy/coordinates/frames.py:133
         │           └─ 0.002 wrap_angle  astropy/coordinates/angles.py:652
         │                 [4 frames hidden]  astropy
         └─ 0.002 wrapper  astropy/units/decorators.py:165
               [7 frames hidden]  astropy, <built-in>
```