@pep8speaks 2020-09-10
Hello @dstansby! Thanks for opening this PR. We checked the lines you've touched for [PEP 8](https://www.python.org/dev/peps/pep-0008) issues, and found:

* In the file [`sunpy/map/maputils.py`](https://github.com/sunpy/sunpy/blob/986d0275b38dda3b777b569af2f97a2d07aaf572/sunpy/map/maputils.py):

> [Line 170:101](https://github.com/sunpy/sunpy/blob/986d0275b38dda3b777b569af2f97a2d07aaf572/sunpy/map/maputils.py#L170): [E501](https://duckduckgo.com/?q=pep8%20E501) line too long (106 > 100 characters)