@nabobalis 2020-09-10
What is the rest of the time spent on?