@dstansby 2020-09-10
> contains_solar_center might be able to be fooled if you throw it a really large field of view, like the whole sky?

Shouldn't be fooled, as by definition of the coordinate origin is the center of the Sun.

> Are we just assuming all maps are in HPC in maputils already or is that an assumption you have made in this diff?

That assumption is all over maputils already. Wouldn't be a bad idea to open an issue/PR to validate the map and throw a nice error if it's not HPC tho.