@dstansby 2020-09-21
Thanks for the reviews! I got a bit confused by them all, so I've resolved all the conversations, tidied up @Cadair's latest suggestion, and force pushed away extra commits. If anyone has more reviews, please could you open new inline threads in the code?