@nabobalis 2020-09-22
> Thoughts on backporting this?
> 
> This isn't an API change, so could be backported without affecting users. Are we willing to establish a policy of performance fixes being considered bug fixes and being backported where there is no behaviour changes associated with them?

It was why I did initially put on the backport label.