@Cadair 2020-09-22
(Test failures seem to be unrelated)

I think I am +1 on backporting perf fixes, it seems to me to be pointless to hold them back from people. I am open to counter arguments thought.