@ayshih 2020-09-22
> I think I am +1 on backporting perf fixes, it seems to me to be pointless to hold them back from people. I am open to counter arguments thought.

What if users are depending on the slower performance?!? =P  (I jest, but sometimes a significant performance improvement can break user code, e.g., if there are parallel operations that implicitly depended on timing.)

+1 to backporting